requirements.txt
Django==4.1.6
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
djoser==2.1.0
rest_framework_simplejwt==0.0.2
